Added a rhombicuboctahedronal configuration
The rhombicuboctahedron is nice because it's smaller faces, but more
importantly exibits the same symmetry that BB8 does - so the joints etc.
should be less noticable.

Plan to make pockets in the corners to hold washers, biscuit-joint
style, and make pocket screwholes for the centers of the faces...
possibly only the square faces.

There are 18 squares, roughly 8"x8", and 8 triangles.  It should be
/barely/ printable - full size - on most 8" printers.

I'll try skin it with bb8 texture, but I haven't found a good
full-sphere model.