Added two washer slots per side.
In order to make it accessible for charging, I plan to not put any washers in the triangles; that way they'll be easily removable (3 screws).

The washers are sized for M5 washers... because I have a ton of them lying around.

I feel weird that the screws are M4 and the washers are M5.  I console myself because M5 washers are 10mm diameter, and ten is a nice round number.

Paul