Endcaps, motor drive gears and motor carriers are done.
Will print a test set soon - waiting on parts for my big printer.  Angles for the texture need to be looked at first.  Next step will be the tilt mechanism, and mocking up a spacer for the head mechanism.

The axle design is less elegant than my hamster :-(  But at least modular enough to get a ball-rolling test before designing the entire thing.

Electronics are behind, so I'll be working on remote control while this set is printing.

Paul