Should be completed controller code.  Not tested, even a little.
I've got some time to code without hardware, so hey.  Go Cubs.
Going to work on the body.

Right now the test bed is just four omniwheel motors, so plan to
get the motion going.

Haven't figured out how to monitor the battery yet... I mean, just
some voltage dividers and an analog in; but how to report it to
the controller?